+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog 1.0.0].
 ### Feat
 
 - **Identifier**: add the ability to store and retrieve if an identifier is deprecated
+- **Document**: add case number and category to the document body if available in the xml
 
 ### Fix